pyproject.toml
[tool.poetry]
name = "aws-idp-receipts-demo"
version = "0.0.1"
description = "DEMO for a receipt IDP platform on top of AWS"
authors = [
    "Santiago Garcia Arango <santigrc@amazon.com>",
    "Daniel Ayala Marin <daniayma@amazon.com>",
]
license = "Apache"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.177.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.0"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
coverage = "^7.4.0"
black = "^23.12.1"
boto3 = "^1.34.14"
aws-lambda-powertools = { version = "^2.31.0" }
fastapi = { extras = ["all"], version = "^0.109.0" }
mangum = "^0.17.0"
pydantic = "^2.5.3"
moto = "^5.0.11"
requests = "^2.32.3"
python-ulid = "^2.2.0"
fpdf2 = "^2.8.2"


[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = ["backend", "cdk"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
black-format = "black ."
black-check = "black . --check --diff -v"
test-unit = ["_test_unit", "_coverage_report", "_coverage_html"]
synth = "cdk synth"
_test_unit = "coverage run -m pytest tests/unit"
_coverage_report = "coverage report"
_coverage_html = "coverage html"

[tool.coverage.run]
branch = true
source = ["backend", "cdk"]
omit = ["**/__init__.py", "cdk/app.py"]

[tool.coverage.report]
show_missing = false
